refactor: dedupe telemetry bootstrap and logging
